Show a banner on app startup when feeder models and the list of applications are not available yet
